PowerSpectrumNoises

Documentation for PowerSpectrumNoises.

PowerSpectrumNoises.AbstractNoiseSignal

PowerSpectrumNoises.AbstractPowerSpectrumModel

PowerSpectrumNoises.DensityAnalytic

PowerSpectrumNoises.IsAnalytic

PowerSpectrumNoises.ModelModifier

PowerSpectrumNoises.ModifiedPowerSpectrumModel

PowerSpectrumNoises.NoiseSignal

PowerSpectrumNoises.NotAnalytic

PowerSpectrumNoises.Renormalize

PowerSpectrumNoises.Rotate

PowerSpectrumNoises.SinglePowerLaw

PowerSpectrumNoises.Stretch

PowerSpectrumNoises.amplitude_point

PowerSpectrumNoises.basemodel

PowerSpectrumNoises.fourieranalytic

PowerSpectrumNoises.generate_gaussian_noise

PowerSpectrumNoises.modify

PowerSpectrumNoises.posangle

PowerSpectrumNoises.power_point

PowerSpectrumNoises.renormed

PowerSpectrumNoises.rfftsize

PowerSpectrumNoises.rotated

PowerSpectrumNoises.scale_fourier

PowerSpectrumNoises.stretched

PowerSpectrumNoises.transform_ν

PowerSpectrumNoises.unmodified

PowerSpectrumNoises.AbstractNoiseSignal — Type

AbstractNoiseSignal

This is an abstract data type for signal data on which powerspectrum noises will be created. The data type contains information to define the dimensions and sizes of data, and provides various methods.

Mandatory fields

dims::Tuple: dimension of data

Mandatory methods

size(data::AbstractNoiseSignal): returns the size of the signal

sizeof(data::AbstractContinuousNoiseSignal): returns product of signal dimensions

ndims(data::AbstractNoiseSignal): returns the number of the dimension

rfftsize(data::AbstractNoiseSignal): returns the size of the signal in the Fourier plane (for rfft)

rfftfreq(data::AbstractNoiseSignal): returns the frequency grid along each dimension in the Fourier plane (for rfft)

generate_gaussian_noise(data::AbstractNoiseSignal): returns Gaussian noises in Fourier domain with the size of rfftsize(data)

PowerSpectrumNoises.AbstractPowerSpectrumModel — Type

AbstractPowerSpectrumModel{N}

The abstract model type for a PowerSpectrum at N dimension. To instantiate your own model type you should subtybe from this model. Additionally you need to implement the following methods to satify the interface:

Mandatory Methods

fourieranalytic: defines whether the model fourier coefficients can be computed analytically. If yes then this should return IsAnalytic() and the user must to define visibility_point. If not analytic then fourieranalytic should return NotAnalytic().

power_point: Defines how to compute model power of fourier coefficients pointwise. Note this is must be defined if fourieranalytic(::Type{YourModel})==IsAnalytic().

Optional Methods

amplitude_point: Defines how to compute model amplitudes of fourier coefficients pointwise. Note this is must be defined if fourieranalytic(::Type{YourModel})==IsAnalytic(). It should be defined as √(power_point(model, ν...)).

PowerSpectrumNoises.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

PowerSpectrumNoises.IsAnalytic — Type

struct IsAnalytic <: PowerSpectrumNoises.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

PowerSpectrumNoises.ModelModifier — Type

abstract type ModelModifier{T}

General type for a model modifier. These transform any model using simple Fourier transform properties. To modify a model you can use the ModifiedPowerSpectrumModel{N} constructor or the modify function.

julia> visanalytic(stretched(Disk(), 2.0, 2.0)) == visanalytic(Disk())
true

To implement a model transform you need to specify the following methods:

transform_ν

transform_image

scale_fourier

scale_image

radialextent

See thee docstrings of those methods for guidance on implementation details.

Additionally these methods assume the modifiers are of the form

I(x,y) -> fᵢ(x,y)I(gᵢ(x,y)) V(u,v) -> fᵥ(u,v)V(gᵥ(u,v))

where g are the transformimage/uv functions and f are the scaleimage/uv function.

PowerSpectrumNoises.ModifiedPowerSpectrumModel — Type

struct ModifiedPowerSpectrumModel{N, M<:AbstractPowerSpectrumModel{N}, T<:Tuple} <: AbstractPowerSpectrumModel{N}

Container type for models that have been transformed in some way. For a list of potential modifiers or transforms see subtypes(ModelModifiers).

Fields

model: base model

transform: model transforms

PowerSpectrumNoises.NoiseSignal — Type

NoiseSignal <: AbstractNoiseSignal

This is a data type for N dimensional noise signals.

PowerSpectrumNoises.NotAnalytic — Type

struct NotAnalytic <: PowerSpectrumNoises.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

PowerSpectrumNoises.Renormalize — Type

struct Renormalize{T} <: PowerSpectrumNoises.ModelModifier{T}

Renormalizes the flux of the model to the new value scale*flux(model). We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

Example

julia> modify(Gaussian(), Renormalize(2.0)) == 2.0*Gaussian()
true

PowerSpectrumNoises.Rotate — Type

Rotate(ξ)

Type for the rotated model. This is more fine grained constrol of rotated model.

Example

julia> modify(Gaussian(), Rotate(2.0)) == rotated(Gaussian(), 2.0)
true

PowerSpectrumNoises.SinglePowerLaw — Type

SinglePowerLaw <: AbstractPowerSpectrumModel

This is a data type for an N dimensional power law function of the form P(ν)=ν^(index)

PowerSpectrumNoises.Stretch — Type

Stretch(α, β)
Stretch(r)

Stretched the model in the x and y directions, i.e. the new intensity is Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

Example

julia> modify(Gaussian(), Stretch(2.0)) == stretched(Gaussian(), 2.0, 1.0)
true

If only a single argument is given it assumes the same stretch is applied in both direction.

julia> Stretch(2.0) == Stretch(2.0, 2.0)
true

PowerSpectrumNoises.amplitude_point — Method

amplitude_point(model::AbstractPowerSpectrumModel, ν...)

Function that computes the pointwise amplitude of fourier component at a set of frequencies ν. This must be implemented in the model interface if fourieranalytic(::Type{MyModel}) == IsAnalytic().

In default, it should be defined as √(power_point(model, ν...)).

PowerSpectrumNoises.basemodel — Method

basemodel(model::ModifiedPowerSpectrumModel)

Returns the ModifiedPowerSpectrumModel with the last transformation stripped.

Example

julia> basemodel(stretched(Disk(), 1.0, 2.0)) == Disk()
true

PowerSpectrumNoises.fourieranalytic — Method

fourieranalytic(::Type{<:AbstractPowerSpectrumModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

PowerSpectrumNoises.generate_gaussian_noise — Method

generate_gaussian_noise(data::AbstractNoiseSignal; rng=Random.default_rng())

Generate complex gausisian noises on Fourier space over the frequency space expected for Real FFT. This return a complex array, which is consistent with Fourier-domain representation of a real uncorrelated Gaussian noise with the zero mean and variance of sizeof(data)/2 in signal-domain. The size of the output array is given by rfftsize(data).

PowerSpectrumNoises.modify — Method

modify(m::AbstractPowerSpectrumModel, transforms...)

Modify a given model using the set of transforms. This is the most general function that allows you to apply a sequence of model transformation for example

modify(Gaussian(), Stretch(2.0, 1.0), Rotate(π/4), Shift(1.0, 2.0), Renorm(2.0))

will create a asymmetric Gaussian with position angle π/4 shifted to the position (1.0, 2.0) with a flux of 2 Jy. This is similar to Flux's chain.

PowerSpectrumNoises.posangle — Method

posangle(model)

Returns the rotation angle of the rotated model

PowerSpectrumNoises.power_point — Function

power_point(model::AbstractPowerSpectrumModel, ν...)

Function that computes the pointwise amplitude of fourier component at a set of frequencies ν. This must be implemented in the model interface if fourieranalytic(::Type{MyModel}) == IsAnalytic()

PowerSpectrumNoises.renormed — Method

renormed(model, f)

Renormalizes the model m to have total flux f*flux(m). This can also be done directly by calling Base.:* i.e.,

julia> renormed(m, f) == f*M
true

PowerSpectrumNoises.rfftsize — Method

rfftsize(data::AbstractNoiseSignal)::Tuple

Returns the size of the Fourier-domain signal

PowerSpectrumNoises.rotated — Method

rotated(model, ξ)

Rotates the model by an amount ξ in radians in the clockwise direction.

PowerSpectrumNoises.scale_fourier — Function

scale_fourier(model::AbstractModifier, ν...)

Returns a number on how to scale the fourier coefficients at the frequency coordinate ν for an modified model

PowerSpectrumNoises.stretched — Method

stretched(model, α)

Stretches the model m according to the formula Vₛ(u) = V(u1/α1, u2/α2, ...), where were renormalize the intensity to preserve the models flux.

PowerSpectrumNoises.transform_ν — Function

transform_ν(model::AbstractModifier, ν...)

Returns a transformed frequency coordinate ν according to the model modifier

PowerSpectrumNoises.unmodified — Method

unmodified(model::ModifiedPowerSpectrumModel)

Returns the un-modified model

Example

julia> m = stretched(rotated(Gaussian(), π/4), 2.0, 1.0)
julia> umodified(m) == Gaussian()
true