PowerSpectrumNoises

Documentation for PowerSpectrumNoises.

PowerSpectrumNoises.AbstractNoisePowerSpectrum

PowerSpectrumNoises.AbstractNoiseSignal

PowerSpectrumNoises.NoiseSignal1D

PowerSpectrumNoises.NoiseSignal2D

PowerSpectrumNoises.generate_gaussian_noise

PowerSpectrumNoises.rfftsize

PowerSpectrumNoises.AbstractNoisePowerSpectrum — Type

AbstractNoisePowerSpectrum

This is an abstract data type for noise power spectra.

Mandatory fields

amp::Float64: amplitude of power spectrum

index::Float64: power law index β

Mandatory methods

-map_ampspectrum_point(psmodel::AbstractNoisePowerSpectrum, singleν::Number): maps power spectrum amplitude for a single frequency point -map_ampspectrum(psmodel::AbstractNoisePowerSpectrum, gridofν::AbstractArray...): maps power spectrum amplitudes for a grid of frequencies -map_powerspectrum_point(psmodel::AbstractNoisePowerSpectrum, singleν::Number): maps power spectrum for a single frequency point -map_powerspectrum(psmodel::AbstractNoisePowerSpectrum, gridofν::AbstractArray...): maps power spectrum for a grid of frequencies

PowerSpectrumNoises.AbstractNoiseSignal — Type

AbstractNoiseSignal

This is an abstract data type for signal data on which powerspectrum noises will be created. The data type contains information to define the dimensions and sizes of data, and provides various methods.

Mandatory fields

dims::Tuple: dimension of data

pixelsizes::Tuple: pixel size along with each dimension

Mandatory methods

size(data::AbstractNoiseSignal): returns the size of the signal

ndims(data::AbstractNoiseSignal): returns the number of the dimension

signalgrid(data::AbstractNoiseSignal): returns position grid in the image plane

rfftsize(data::AbstractNoiseSignal): returns the size of the signal in the Fourier plane (for rfft)

rfftfreq(data::AbstractNoiseSignal): returns the frequency grid along each dimension in the Fourier plane (for rfft)

generate_gaussian_noise(data::AbstractNoiseSignal): returns Gaussian noises in Fourier domain with the size of rfftsize(data)

PowerSpectrumNoises.NoiseSignal1D — Type

NoiseSignal1D <: AbstractNoiseSignal

This is a data type for one dimensional noise signals.

PowerSpectrumNoises.NoiseSignal2D — Type

NoiseSignal2D <: AbstractNoiseSignal

This is a data type for two dimensional noise signals.

Mandatory methods

freq_norm: 2D matrix of norms of each frequency point in fourier space

PowerSpectrumNoises.generate_gaussian_noise — Method

generate_gaussian_noise(data::AbstractNoiseSignal; rng=Random.default_rng())

Generate complex gausisian noises on Fourier space over the frequency space expected for Real FFT. This return a complex array, which is consistent with Fourier-domain representation of a real uncorrelated Gaussian noise with the zero mean and variance of sizeof(data)/2 in signal-domain. The size of the output array is given by rfftsize(data).

PowerSpectrumNoises.rfftsize — Method

rfftsize(data::AbstractNoiseSignal)::Tuple

Returns the size of the Fourier-domain signal